Refactor `@/notifier/model/notificationEvents`
