+++ packages/server/src/notifier/model/notificationEvents/fromEvents/index.ts
@@ -1,0 +1,1 @@
+export * from './forum'
